+++ ai21/clients/common/beta/assistant/threads.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 from abc import ABC, abstractmethod
 from typing import List
 
-from ai21.clients.common.assistant.messages import Messages
+from ai21.clients.common.beta.assistant.messages import Messages
 from ai21.models.assistant.message import Message
 from ai21.models.responses.thread_response import ThreadResponse
