+++ ai21/clients/studio/resources/beta/assistant/thread_message.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 
-from ai21.clients.common.assistant.messages import Messages
+from ai21.clients.common.beta.assistant.messages import Messages
 from ai21.clients.studio.resources.studio_resource import StudioResource, AsyncStudioResource
 from ai21.models.assistant.message import ThreadMessageRole, MessageContentText
 from ai21.models.responses.message_response import MessageResponse, ListMessageResponse
